Move deserialize inside proxy
