+++ utils/interfaces/_core.py
@@ -10,7 +10,6 @@ import re
 import time
 
 from utils.tools import logger
-from ._deserializer import deserialize
 
 
 class InterfaceValidator:
@@ -65,11 +64,6 @@ class InterfaceValidator:
                 except json.decoder.JSONDecodeError:
                     # the file may not be finished
                     return
-
-            deserialize(data, self.name)
-
-            with open(src_path, "w", encoding="utf-8") as f:
-                json.dump(data, f, indent=2, cls=ObjectDumpEncoder)
 
             self._data_list.append(data)
             self._ingested_files.add(src_path)
@@ -146,13 +140,6 @@ class InterfaceValidator:
             logger.error(f"Wait for {wait_for_function} finished in error")
 
         self._wait_for_function = None
-
-
-class ObjectDumpEncoder(json.JSONEncoder):
-    def default(self, o):
-        if isinstance(o, bytes):
-            return str(o)
-        return json.JSONEncoder.default(self, o)
 
 
 class ValidationError(Exception):
