+++ utils/proxy/Dockerfile
@@ -1,0 +1,3 @@
+FROM mitmproxy/mitmproxy:9.0.1
+
+RUN pip install requests-toolbelt==0.10.1 grpcio-tools==1.56.0 opentelemetry-proto==1.17.0
