+++ utils/proxy/_decoders/trace_payload.descriptor
